Notice resolution
If PHP is set to give a notice when an unset variable is used, changes needed to be made to resolve those.

+++ cpanel/page_edit.php
@@ -1,7 +1,7 @@
 <?php
 require_once("load_page.php");
 
-if(isset($_GET['confirmdelete']) && is_numeric($_GET['pageId']))
+if(isset($_GET['confirmdelete']) && !empty($_GET['pageId']) && is_numeric($_GET['pageId']))
 {
 	if($builder->database->delete("pages", ['index'=>(int)$_GET['pageId']], true))
 		$_SESSION['onload_notification'] = [['__attr:type'=>"primary", "Page was deleted successfully."], '__attr:title'=>"Page Deleted"];
@@ -11,15 +11,15 @@ if(isset($_GET['confirmdelete']) && is_numeric($_GET['pageId']))
 	exit;
 }
 
-if(is_numeric($_GET['pageId']))
+if(!empty($_GET['pageId']) && is_numeric($_GET['pageId']))
 	$page = $builder->database->query("SELECT * FROM `pages` WHERE `index`=". (int)$_GET['pageId'] ." LIMIT 0,1", 2);
-else if(is_numeric($_GET['specialPageId']))
+else if(!empty($_GET['specialPageId']) && is_numeric($_GET['specialPageId']))
 {
 	$special = true;
 	$page = $builder->database->query("SELECT * FROM `pages_special` WHERE `index`=". (int)$_GET['specialPageId'] ." LIMIT 0,1", 2);
 }
 
-if(is_array($page) || $_GET['pageId'] == "new")
+if(!empty($page) && is_array($page) || !empty($_GET['pageId']) && $_GET['pageId'] == "new")
 {
 	$builder->add_content(new Text(['component'=>$builder->database->query("SELECT `index`,`title` FROM `page_components` ORDER BY `title`", Database::RETURN_ALL)], ['hidden'=>"1"]), "component-list");
 	$dbtables = [];
@@ -42,20 +42,20 @@ if(is_array($page) || $_GET['pageId'] == "new")
 	$builder->add_content(new Text(['class'=>Content::get_subclasses($builder)], ['hidden'=>"1"]), "content-classes");
 	$form = $builder->add_content(new Container("", []), "page_edit");
 	$data = [
-		'title' => $page['title'],
+		'title' => empty($page['title']) ? "" : $page['title'],
 		'site_title' => $builder->get_setting('site_title'),
-		'select@id=subtheme' => ['value'=>$page['subtheme'], 'option'=>[]],
+		'select@id=subtheme' => ['value'=>empty($page['subtheme'])?"":$page['subtheme'], 'option'=>[]],
 		'select@id=page_css@multiple' => ['value'=>[],'option'=>[]],
 		'select@id=page_js@multiple' => ['value'=>[],'option'=>[]],
 		'select@id=page_xsl@multiple' => ['value'=>[],'option'=>[]],
 	];
 	foreach($builder->themes[$builder->get_theme()]['subtheme'] as $subtheme)
 		$data['select@id=subtheme']['option'][] = $subtheme['__attr:name'];
-	if(!$special)
+	if(empty($special))
 	{
-		if($page['index'])
+		if(!empty($page['index']))
 			$data['index'] = $page['index'];
-		$data['url'] = $page['url'];
+		$data['url'] = empty($page['url']) ? "" : $page['url'];
 		$data['url_path'] = $builder->get_setting('url_path');
 		$data['select@id=permissions@multiple'] = ['value'=>[],'option'=>[]];
 	}
@@ -75,7 +75,7 @@ if(is_array($page) || $_GET['pageId'] == "new")
 				$data['select@id=page_css@multiple']['option'][] = [$entry];
 			}
 	}
-	$page_css = json_decode($page['css'], true);
+	$page_css = json_decode(empty($page['css'])?"":$page['css'], true);
 	if(is_array($page_css)) foreach($page_css as $css)
 	{
 		$data['select@id=page_css@multiple']['value'][] = $css['file'];
@@ -91,7 +91,7 @@ if(is_array($page) || $_GET['pageId'] == "new")
 				$data['select@id=page_js@multiple']['option'][] = [$entry];
 			}
 	}
-	$page_js = json_decode($page['js'], true);
+	$page_js = json_decode(empty($page['js'])?"":$page['js'], true);
 	if(is_array($page_js)) foreach($page_js as $js)
 	{
 		$data['select@id=page_js@multiple']['value'][] = $js['file'];
@@ -107,18 +107,18 @@ if(is_array($page) || $_GET['pageId'] == "new")
 				$data['select@id=page_xsl@multiple']['option'][] = [$entry];
 			}
 	}
-	$page_xsl = json_decode($page['xsl'], true);
+	$page_xsl = json_decode(empty($page['xsl'])?"":$page['xsl'], true);
 	if(is_array($page_xsl)) foreach($page_xsl as $xsl)
 	{
 		$data['select@id=page_xsl@multiple']['value'][] = $xsl;
 	}
 	
-	if(!$special)
+	if(empty($special))
 	{
 		foreach(User::get_permissions($builder) as $num=>$perm)
 		{
 			$data['select@id=permissions@multiple']['option'][] = [$perm, "__attr:value"=>$num];
-			if(($num & $page['permission']) == $num)
+			if(($num & (empty($page['permission'])?0:$page['permission'])) == $num)
 				$data['select@id=permissions@multiple']['value'][] = $num;
 		}
 	}
@@ -160,14 +160,14 @@ if(is_array($page) || $_GET['pageId'] == "new")
 		}
 		return ['content'=>$result];
 	}
-	$page_contents = $form->add_content(new Text(getContentData(unserialize($page['content']))), "page_content");
+	$page_contents = $form->add_content(new Text( empty($page['content']) ? "" : getContentData(unserialize($page['content'])) ), "page_content");
 	//$page_contents->add_content(getContentData(unserialize($page['content'])));
 
 	$builder->attach_xsl("cpanel-page-edit.xsl", "", true);
 	$builder->attach_xsl("cpanel-content.xsl", "", true);
 	$builder->attach_js("cpanel-page-content.js", "", true);
 	$builder->attach_css("cpanel.css", "", true);
-	$builder->render($_GET['output']=="xml"?"__xml":"cpanel");
+	$builder->render((!empty($_GET['output']) && $_GET['output']=="xml") ? "__xml" : "default");
 }
 else
 {
